MS4/baseline/requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
array-record==0.2.0
asgiref==3.6.0
astunparse==1.6.3
async-generator==1.10
async-timeout==4.0.2
attrs==22.2.0
beautifulsoup4==4.11.2
blinker==1.6.2
bs4==0.0.1
cachetools==5.3.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==2.2.1
colorama==0.4.6
contourpy==1.0.6
cycler==0.11.0
datasets==2.13.1
DateTime==5.1
dill==0.3.6
Django==3.0.1
dm-tree==0.1.8
etils==1.3.0
exceptiongroup==1.1.1
fake-useragent==1.1.1
filelock==3.12.2
Flask==2.3.2
flatbuffers==23.5.26
fonttools==4.38.0
frozenlist==1.4.0
fsspec==2023.6.0
gast==0.4.0
gensim==4.3.1
gevent==22.10.2
google-auth==2.19.1
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
googleapis-common-protos==1.59.1
greenlet==2.0.2
grequests==0.6.0
grpcio==1.54.2
h11==0.14.0
h5py==3.8.0
hmmlearn==0.3.0
huggingface-hub==0.15.1
idna==3.4
image==1.5.33
imageio==2.31.0
imbalanced-learn==0.10.1
imblearn==0.0
importlib-resources==5.12.0
iniconfig==2.0.0
itsdangerous==2.1.2
jax==0.4.12
Jinja2==3.1.2
jiwer==3.0.2
joblib==1.2.0
keras==2.12.0
kiwisolver==1.4.4
lazy_loader==0.2
libclang==16.0.0
lime==0.2.0.1
llvmlite==0.40.1rc1
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
ml-dtypes==0.2.0
multidict==6.0.4
multiprocess==0.70.14
mysql==0.0.3
mysql-connector==2.2.9
mysql-connector-python==8.0.32
mysqlclient==2.1.1
networkx==3.1
numba==0.57.0
numpy==1.23.5
oauthlib==3.2.2
opencv-python==4.7.0.72
opt-einsum==3.3.0
outcome==1.2.0
packaging==21.3
panda==0.3.1
pandas==1.5.3
patsy==0.5.3
Pillow==9.3.0
pluggy==1.0.0
promise==2.3
protobuf==3.20.3
psutil==5.9.5
pyarrow==12.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyMySQL==1.0.2
pyparsing==3.0.9
PySocks==1.7.1
pytest==7.2.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7.1
PyWavelets==1.4.1
PyYAML==6.0
rapidfuzz==2.13.7
regex==2022.10.31
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
safetensors==0.3.1
scikit-image==0.21.0
scikit-learn==1.2.2
scikit-surprise==1.1.3
scipy==1.10.1
selenium==4.8.2
shap==0.41.0
six==1.16.0
sklearn==0.0
slicer==0.0.7
smart-open==6.3.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4
sqlparse==0.4.3
statsmodels==0.14.0
surprise==0.1
tensorboard==2.12.3
tensorboard-data-server==0.7.0
tensorflow==2.12.0
tensorflow-datasets==4.9.2
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.32.0
tensorflow-metadata==1.13.1
termcolor==2.3.0
textparser==0.24.0
threadpoolctl==3.1.0
tifffile==2023.4.12
tokenizers==0.13.3
toml==0.10.2
tqdm==4.65.0
transformers==4.30.2
trio==0.22.0
trio-websocket==0.10.0
typing_extensions==4.6.3
tzdata==2022.7
urllib3==1.26.15
vc==2018.7.10
webdriver-manager==3.8.5
Werkzeug==2.3.4
wrapt==1.14.1
wsproto==1.2.0
xxhash==3.2.0
yarl==1.9.2
zipp==3.15.0
zope.event==4.6
zope.interface==5.5.2
